Fix MSVC compiler warnings for Windows CI
- Fixed size_t to int conversions with static_cast
- Resolved variable shadowing in spawn_manager (renamed rng parameter)
- Removed unused exception variables in catch blocks
- Fixed double to float conversion in pathfinding
- Fixed potential unsigned underflow in loop conditions
- Keep _CRT_SECURE_NO_WARNINGS for standard C functions

All warnings fixed through proper code changes rather than suppressions.
This ensures Windows CI builds successfully with /WX (warnings as errors).

+++ include/inventory.h
@@ -131,7 +131,7 @@ public:
     bool hasSpace(int count = 1) const;
 
     /** @brief Get number of occupied slots @return Used slot count */
-    int getUsedSlots() const { return items.size(); }
+    int getUsedSlots() const { return static_cast<int>(items.size()); }
 
     /** @brief Get total slot capacity @return Maximum slots */
     int getTotalSlots() const { return capacity; }
